+++ src/shared/components/case-editor/services/cases.service.ts
@@ -1,4 +1,4 @@
-import { HttpHeaders } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { plainToClass } from 'class-transformer';
 import { Observable, throwError } from 'rxjs';
@@ -54,6 +54,22 @@ export class CasesService {
    * @deprecated Use `CasesService::getCaseView` instead
    */
   get = this.getCaseView;
+
+  public static updateChallengedAccessRequestAttributes(httpClient: HttpClient, caseId: string, attributesToUpdate: { [x: string]: any })
+    : Observable<RoleAssignmentResponse> {
+    return httpClient.post<RoleAssignmentResponse>(`/api/challenged-access-request/update-attributes`, {
+      caseId,
+      attributesToUpdate
+    });
+  }
+
+  public static updateSpecificAccessRequestAttributes(httpClient: HttpClient, caseId: string, attributesToUpdate: { [x: string]: any })
+    : Observable<RoleAssignmentResponse> {
+    return httpClient.post<RoleAssignmentResponse>(`/api/specific-access-request/update-attributes`, {
+      caseId,
+      attributesToUpdate
+    });
+  }
 
   constructor(
     private http: HttpService,
@@ -334,7 +350,7 @@ export class CasesService {
   }
 
   public getCourtOrHearingCentreName(locationId: number): Observable<any> {
-    return this.http.get(`${this.appConfig.getLocationRefApiUrl()}/building-locations?epimms_id=${locationId}`);
+    return this.http.post(`/api/locations/getLocationsById`, { locations : [{ locationId : locationId }]});
   }
 
   public createChallengedAccessRequest(caseId: string, request: ChallengedAccessRequest): Observable<RoleAssignmentResponse> {
@@ -352,6 +368,8 @@ export class CasesService {
     const beginTime = new Date();
     const endTime = new Date(new Date().setUTCHours(23, 59, 59, 999));
     const id = userInfo.id ? userInfo.id : userInfo.uid;
+    const isNew = true;
+
     const payload: RoleRequestPayload = camUtils.getAMPayload(id,
                                                               id,
                                                               roleName,
@@ -360,7 +378,9 @@ export class CasesService {
                                                               caseId,
                                                               request,
                                                               beginTime,
-                                                              endTime);
+                                                              endTime,
+                                                              isNew
+      );
 
     return this.http.post(`/api/challenged-access-request`, payload);
   }
@@ -379,7 +399,7 @@ export class CasesService {
     const roleName = camUtils.getAMRoleName('specific', roleCategory);
     const id = userInfo.id ? userInfo.id : userInfo.uid;
     const payload: RoleRequestPayload = camUtils.getAMPayload(null, id,
-                                      roleName, roleCategory, 'SPECIFIC', caseId, sar);
+                                      roleName, roleCategory, 'SPECIFIC', caseId, sar, null, null, true);
 
     payload.roleRequest = {
       ...payload.roleRequest,
@@ -402,17 +422,18 @@ export class CasesService {
 
     payload.requestedRoles[0].attributes = {
       ...payload.requestedRoles[0].attributes,
-      requestedRole: roleName
+      requestedRole: roleName,
+      specificAccessReason: sar.specificReason
     }
 
     payload.requestedRoles[0].notes[0] = {
       ...payload.requestedRoles[0].notes[0],
       userId: payload.requestedRoles[0].actorId
     }
+
     return this.http.post(
       `/api/specific-access-request`,
       payload
     );
   }
-
 }