+++ src/shared/components/palette/index.ts
@@ -24,3 +24,4 @@ export * from './unsupported-field.component';
 export * from './utils';
 export * from './waystopay';
 export * from './yes-no';
+export * from './judicial-user';