+++ RELEASE-NOTES.md
@@ -1,7 +1,10 @@
 ## RELEASE NOTES
 
-### Version 4.21.0-no-cases-to-unlink-error-message
-**EUI-7234** Linked cases no cases to unlink error message
+### Version 4.22.0-judicial-user-typeahead
+**EUI-6746** CR-52 JOH typeahead
+
+### Version 4.18.16-EUI-6309-mandatory-field-issue
+**EUI-6309** mandatory-field-issue
 
 ### Version 4.19.3-hearings
 **EUI-5912** Hearing tab is not loading